server/requirements.txt
websockets>=11.0
pyserial>=3.5
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sounddevice>=0.4.6
numpy>=1.24.0
uvloop>=0.17.0
